Begin building out file watcher, separate RBX stuff from virtual filesystem
