+++ server/src/lib.rs
@@ -12,10 +12,13 @@ extern crate tempfile;
 
 pub mod commands;
 pub mod file_route;
+pub mod fs_session;
 pub mod id;
 pub mod partition;
 pub mod pathext;
 pub mod project;
-pub mod rbx_session;
+pub mod rbx;
+// pub mod session;
+pub mod session_config;
 pub mod web;
 pub mod web_util;
