+++ server/src/session_config.rs
@@ -1,0 +1,8 @@
+use std::collections::HashMap;
+
+use partition::Partition;
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct SessionConfig {
+    pub partitions: HashMap<String, Partition>,
+}
